+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 3.0.0 (TBD)
+
+### Additions
+
+* Support for graph based models.
+
+### Breaking Changes
+
+* The `to_model` method on `Dbee::Base` objects has been removed. Use `to_schema` instead.
 # 2.1.1 (July 14th, 2020)
 
 * Removed guard that ensured a query has at least one field to establish a more rational base-case.